resolve code review feedback
