+++ azure-spring-boot/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.env.EnvironmentPostProcessor=com.microsoft.azure.spring.cloundfoundry.environment.VcapProcessor,\
-com.microsoft.azure.keyvault.spring.KeyVaultProcessor
+com.microsoft.azure.keyvault.spring.KeyVaultEnvironmentPostProcessor
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.microsoft.azure.spring.autoconfigure.documentdb.DocumentDBAutoConfiguration,\
 com.microsoft.azure.spring.autoconfigure.documentdb.DocumentDbRepositoriesAutoConfiguration,\
 com.microsoft.azure.spring.autoconfigure.mediaservices.MediaServicesAutoConfiguration,\
